F [#4ubrLq42]: Extract coin treasure generation to Treasure::Coins class
